app: 9 reduce code
