+++ app/src/main/java/com/iorlova/diploma/UI/FileExtension.kt
@@ -1,0 +1,24 @@
+package com.iorlova.diploma.UI
+
+import android.content.ContentResolver
+import android.net.Uri
+import android.os.Environment
+import java.io.File
+import java.io.InputStream
+
+fun createNewFile(bookUri: Uri, contentResolver: ContentResolver): File {
+    val file = File(Environment.getExternalStorageDirectory().toString() + "/" + File.separator + "tmpBook")
+    val bookInputStream = contentResolver.openInputStream(bookUri)
+
+    file.createNewFile()
+    file.copyInputStreamToFile(bookInputStream!!)
+    file.deleteOnExit()
+
+    return file
+}
+
+private fun File.copyInputStreamToFile(inputStream: InputStream) {
+    this.outputStream().use { fileOut ->
+        inputStream.copyTo(fileOut)
+    }
+}
